Json (#16)
* add logger to sbml.py
* add option for JSON output
* add more Python versions to ci tests
* enable the use of python -m meneco
* remove old test dir
* quit without reading the repair file if no unproducible target
* simplify import for example, add links to papers

+++ meneco/query.py
@@ -22,84 +22,21 @@ from clyngor.as_pyasp import TermSet, Atom
 from meneco import utils
 
 
-root                      = __file__.rsplit('/', 1)[0]
-unproducible_prg          = root + '/encodings/unproducible_targets.lp'
-ireaction_prg             = root + '/encodings/ireactions.lp'
-minimal_completion_prg    = root + '/encodings/card_min_completions_all_targets.lp'
-heuristic_prg             = root + '/encodings/heuristic.lp'
-minimal_completion_wb_prg = root + '/encodings/card_min_completions_all_targets_with_bounds.lp'
-completion_prg            = root + '/encodings/completions_all_targets.lp'
-
-
-# def get_mapping_ireaction(termset):
-#     dict    = {}
-#     revdict = {}
-#     for a in termset:
-#       if a.pred() == "ireaction" :
-#         if not a.arg(0) in dict:
-#           id             = len(dict)
-#           dict[a.arg(0)] = id
-#           revdict[id]    = a.arg(0)
-#     return dict, revdict
-
-
-# def map_reaction_ids(termset, dict):
-#     mapped = TermSet()
-#     for a in termset:
-#       if a.pred() == "reaction" :
-#         if a.arg(0) in dict:
-#           mapped.add(Term('reaction', [str(dict[a.arg(0)]), a.arg(1)]))
-#         else : mapped.add(a)
-
-#       elif a.pred() == "xreaction" :
-#         if a.arg(0) in dict:
-#           mapped.add(Term('xreaction', [str(dict[a.arg(0)]), a.arg(1)]))
-#         else : mapped.add(a)
-
-#       elif a.pred() == "ireaction" :
-#         if a.arg(0) in dict:
-#           mapped.add(Term('ireaction', [str(dict[a.arg(0)]), a.arg(1)]))
-#         else : print("Error: unknown ireaction, query.py line 64")
-
-#       elif a.pred() == "value" :
-#         if a.arg(0) in dict:
-#           mapped.add(Term('value', [str(dict[a.arg(0)]), a.arg(1)]))
-#         else : mapped.add(a)
-
-#       elif a.pred() == "product" :
-#         if a.arg(1) in dict:
-#           mapped.add(Term('product', [a.arg(0), str(dict[a.arg(1)]),a.arg(2)]))
-#         else : mapped.add(a)
-
-#       elif a.pred() == "reactant" :
-#         if a.arg(1) in dict:
-#           mapped.add(Term('reactant', [a.arg(0), str(dict[a.arg(1)]),a.arg(2)]))
-#         else : mapped.add(a)
-
-#       elif a.pred() == "reversible" :
-#         if a.arg(0) in dict:
-#           mapped.add(Term('reversible', [str(dict[a.arg(0)])]))
-#         else : mapped.add(a)
-#       else :
-#         mapped.add(a)
-
-#     return mapped
-
-
-# def unmap_reaction_ids(termset, revdict):
-#     unmapped = TermSet()
-#     for a in termset:
-#       if a.pred() == "xreaction" :
-#         unmapped.add(Term('xreaction', [str(revdict[int(a.arg(0))]), a.arg(1)]))
-
-#     return unmapped
+root = __file__.rsplit('/', 1)[0]
+unproducible_prg = root + '/encodings/unproducible_targets.lp'
+ireaction_prg = root + '/encodings/ireactions.lp'
+minimal_completion_prg = root + '/encodings/card_min_completions_all_targets.lp'
+heuristic_prg = root + '/encodings/heuristic.lp'
+minimal_completion_wb_prg = root + \
+    '/encodings/card_min_completions_all_targets_with_bounds.lp'
+completion_prg = root + '/encodings/completions_all_targets.lp'
 
 
 def get_unproducible(draft, seeds, targets):
-    draft_f  = utils.to_file(draft)
-    seed_f   =  utils.to_file(seeds)
+    draft_f = utils.to_file(draft)
+    seed_f = utils.to_file(seeds)
     target_f = utils.to_file(targets)
-    prg      = [unproducible_prg, draft_f, seed_f, target_f ]
+    prg = [unproducible_prg, draft_f, seed_f, target_f]
     options = ''
     best_model = None
     models = clyngor.solve(prg, options=options)
@@ -113,7 +50,7 @@ def get_unproducible(draft, seeds, targets):
 
 def compute_ireactions(instance):
     instance_f = utils.to_file(instance)
-    prg        = [ ireaction_prg, instance_f]
+    prg = [ireaction_prg, instance_f]
     best_model = None
 
     models = clyngor.solve(prg)
@@ -122,44 +59,46 @@ def compute_ireactions(instance):
     os.unlink(instance_f)
 
     output = TermSet()
-    for pred in best_model :
-        if pred == 'ireaction' :
-            for a in best_model[pred] : 
-                output.add(Atom('ireaction(\"' + a[0] +'\",\"' + a[1] + '\")'))
+    for pred in best_model:
+        if pred == 'ireaction':
+            for a in best_model[pred]:
+                output.add(
+                    Atom('ireaction("' + a[0] + '","' + a[1] + '")'))
 
     return output
 
 
 def get_minimal_completion_size(draft, repairnet, seeds, targets):
 
-    instance   = TermSet(draft.union(repairnet).union(targets).union(seeds))
+    instance = TermSet(draft.union(repairnet).union(targets).union(seeds))
     ireactions = compute_ireactions(instance)
-    instance   = TermSet(instance.union(ireactions))
+    instance = TermSet(instance.union(ireactions))
     instance_f = utils.to_file(instance)
 
-    prg        = [minimal_completion_prg, instance_f]
+    prg = [minimal_completion_prg, instance_f]
 
-    co         = "--configuration=jumpy --opt-strategy=usc,5"
+    co = "--configuration=jumpy --opt-strategy=usc,5"
 
     optimum = None
     models = clyngor.solve(prg, options=co)
     for model in models.discard_quotes.by_arity:
         optimum = model
-  
+
     os.unlink(instance_f)
     return optimum
 
 
 def get_intersection_of_optimal_completions(draft, repairnet, seeds, targets, optimum):
 
-    instance   = TermSet(draft.union(repairnet).union(targets).union(seeds))
+    instance = TermSet(draft.union(repairnet).union(targets).union(seeds))
     ireactions = compute_ireactions(instance)
-    instance   = TermSet(instance.union(ireactions))
+    instance = TermSet(instance.union(ireactions))
     instance_f = utils.to_file(instance)
 
-    prg        = [minimal_completion_prg, instance_f]
+    prg = [minimal_completion_prg, instance_f]
 
-    options = '--configuration=jumpy --opt-strategy=usc,5 --enum-mode=cautious --opt-mode=optN,' + str(optimum)
+    options = '--configuration=jumpy --opt-strategy=usc,5 --enum-mode=cautious --opt-mode=optN,' + \
+        str(optimum)
     best_model = None
     models = clyngor.solve(prg, options=options)
     for model in models.discard_quotes.by_arity:
@@ -171,13 +110,14 @@ def get_intersection_of_optimal_completions(draft, repairnet, seeds, targets, op
 
 def get_union_of_optimal_completions(draft, repairnet, seeds, targets, optimum):
 
-    instance   = TermSet(draft.union(repairnet).union(targets).union(seeds))
+    instance = TermSet(draft.union(repairnet).union(targets).union(seeds))
     ireactions = compute_ireactions(instance)
-    instance   = TermSet(instance.union(ireactions))
+    instance = TermSet(instance.union(ireactions))
     instance_f = utils.to_file(instance)
-    prg        = [minimal_completion_prg, instance_f]
+    prg = [minimal_completion_prg, instance_f]
 
-    options = '--configuration jumpy --opt-strategy=usc,5 --enum-mode=brave --opt-mode=optN,' + str(optimum)
+    options = '--configuration jumpy --opt-strategy=usc,5 --enum-mode=brave --opt-mode=optN,' + \
+        str(optimum)
 
     models = clyngor.solve(prg, options=options)
     for model in models.discard_quotes.by_arity:
@@ -187,19 +127,20 @@ def get_union_of_optimal_completions(draft, repairnet, seeds, targets, optimum):
     return best_model
 
 
-
 def get_optimal_completions(draft, repairnet, seeds, targets, optimum, nmodels=0):
 
-    instance   = TermSet(draft.union(repairnet).union(targets).union(seeds))
+    instance = TermSet(draft.union(repairnet).union(targets).union(seeds))
     ireactions = compute_ireactions(instance)
-    instance   = TermSet(instance.union(ireactions))
+    instance = TermSet(instance.union(ireactions))
     instance_f = utils.to_file(instance)
 
-    prg        = [minimal_completion_prg, instance_f]
+    prg = [minimal_completion_prg, instance_f]
 
-    options = '--configuration=handy --opt-strategy=usc,0 --opt-mode=optN,' + str(optimum)
+    options = '--configuration=handy --opt-strategy=usc,0 --opt-mode=optN,{0}'.format(
+        optimum)
 
-    models = clyngor.solve(prg, options=options, nb_model=nmodels).by_arity.discard_quotes
+    models = clyngor.solve(prg, options=options,
+                           nb_model=nmodels).by_arity.discard_quotes
     opt_models = clyngor.opt_models_from_clyngor_answers(models)
 
     return opt_models
@@ -207,19 +148,21 @@ def get_optimal_completions(draft, repairnet, seeds, targets, optimum, nmodels=0
 
 def get_intersection_of_completions(draft, repairnet, seeds, targets):
 
-    instance   = TermSet(draft.union(repairnet).union(targets).union(seeds))
+    instance = TermSet(draft.union(repairnet).union(targets).union(seeds))
     ireactions = compute_ireactions(instance)
-    instance   = TermSet(instance.union(ireactions))
-    instance_f = utils.to_file(instance)
+    if len(ireactions) == 0:
+        return {}
+    else:
+        instance = TermSet(instance.union(ireactions))
+        instance_f = utils.to_file(instance)
 
-    prg        = [completion_prg, instance_f]
-    options    = '--enum-mode=cautious --opt-mode=ignore '
+        prg = [completion_prg, instance_f]
+        options = '--enum-mode=cautious --opt-mode=ignore '
 
-    
-    best_model = None
-    models = clyngor.solve(prg, options=options)
-    for model in models.discard_quotes.by_arity:
-        best_model = model
+        best_model = None
+        models = clyngor.solve(prg, options=options)
+        for model in models.discard_quotes.by_arity:
+            best_model = model
 
-    os.unlink(instance_f)
-    return best_model
+        os.unlink(instance_f)
+        return best_model
